backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.10.3
pydantic-settings==2.6.1
pyotp==2.9.0
cryptography==43.0.3
python-multipart==0.0.17
brotli==1.1.0
